Adds the gui.py and adds the moded classes + new class
The gui.py acts as the interface where users enter their ranges for bedroom / bath numbers, price, and distance from campus and they can select which is prioritized in their search.

The modified classes cuts down the language for the subclasses already present, other group members or myself will go back and ass descriptions/comments for everything that is missing. This also adds in the new class FilterPreference where a user can select whether or not they want to prioritize a specific part of the search program. If they do not prioritize certain things then they can be shown more apartments that may be included in a larger range of filtering.